refactor: create Game components directory
