+++ interfaces/Game/ColorDisplayerProps.ts
@@ -1,4 +1,4 @@
-import ColorInfo from './ColorInfo';
+import ColorInfo from '../ColorInfo';
 
 interface ColorDisplayerProps {
 	/**
