+++ interfaces/Game/ColorRowProps.ts
