+++ interfaces/Game/GuessCellProps.ts
