+++ interfaces/Game/GuessInputProps.ts
